requirements.txt
absl-py==0.12.0
astor==0.8.1
astunparse==1.6.3
autopep8==1.5.6
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.22
dataclasses==0.8
detectron2==0.1.3+cpu
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
fvcore==0.1.5.post20210402
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
iopath==0.1.8
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mock==4.0.3
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.2.0
pkg-resources==0.0.0
portalocker==2.3.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.7.0
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
toml==0.10.2
torch==1.5.0+cpu
torchvision==0.6.0+cpu
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
uuid==1.30
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8
zipp==3.4.1
